pyproject.toml
[project]
name = "conexiones"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-stubs>=5.2.5",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.0",
    "django-unfold>=0.40.0",
]
